[Edit]: WritingView -> RecordView + 뷰간 연결 #81
